requirements.txt
streamlit>=1.28.0
langchain>=0.0.350
langchain-openai>=0.1.0
langchain-community>=0.0.20
PyPDF2>=3.0.1
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
openai>=1.3.0
tiktoken>=0.5.1
sentence-transformers>=2.2.2

